Add Crepe::Application and a config mechanism
Give Crepe a Rails-like application wrapper that allows for dynamic
configuration. This is useful for per-environment settings or simply
"global" accessors to application variables. The Application class will
also be used for future work involved in initializing larger Crepe apps
that use Creperie, such as loading specific environments and
initializers.

Signed-off-by: David Celis <me@davidcel.is>